PointCA_TF/requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1637088766493/work
addict==2.4.0
anyio==3.4.0
argcomplete==1.12.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor @ file:///home/conda/feedstock_root/build_artifacts/astor_1593610464257/work
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi==1.15.0
chamfer-3D==2.1
charset-normalizer==2.0.9
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
Cython==0.29.24
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
deprecation==2.1.0
entrypoints==0.3
filelock==3.3.0
future==0.18.2
gast @ file:///home/conda/feedstock_root/build_artifacts/gast_1636964356021/work
grpcio @ file:///home/conda/feedstock_root/build_artifacts/grpcio_1645230332563/work
h5py==2.9.0
idna==3.3
imageio==2.3.0
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1646003248740/work
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.0.1
json5==0.9.6
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-packaging==0.11.1
jupyter-server==1.13.1
jupyterlab==3.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing @ file:///home/conda/feedstock_root/build_artifacts/keras-preprocessing_1610713559828/work
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1635836699543/work
lmdb==1.0.0
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1637220118004/work
MarkupSafe==2.0.1
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-base_1594091695977/work
matplotlib-inline==0.1.3
mistune==0.8.4
mock @ file:///home/conda/feedstock_root/build_artifacts/mock_1635819543091/work
msgpack==0.5.6
msgpack-numpy==0.4.7.1
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
nose==1.3.7
notebook==6.4.6
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1640083064494/work
open3d==0.13.0
open3d-python==0.7.0.0
opencv-python==3.4.4.19
packaging==21.3
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1592421991297/work
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
Pillow-PIL==0.1.dev0
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.4
psutil==5.8.0
ptyprocess==0.7.0
py==1.11.0
pyarrow==6.0.1
pycparser==2.21
Pygments==2.10.0
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1642753572664/work
pyrsistent==0.18.0
pytest==7.0.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-pcl==0.3.0a1
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1633452062248/work
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
scikit-learn==0.24.2
scipy==1.2.0
seaborn==0.9.0
Send2Trash==1.8.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sklearn==0.0
sniffio==1.2.0
tabulate==0.8.9
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorflow-probability==0.6.0
tensorpack==0.8.9
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tflearn==0.3.2
threadpoolctl==2.2.0
tomli==2.0.1
tomlkit==0.7.2
torch==1.6.0+cpu
torchvision==0.7.0+cpu
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1635819587639/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1610419111160/work
traitlets==5.1.1
transforms3d==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1644850595256/work
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1644332431572/work
widgetsnbextension==3.5.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1643828507773/work
